refactor(llm-auditor): migrate to core/kotlin and adk-kotlin 0.9.0
The only Kotlin recipe still lived under the retired kotlin/agents tree, where
CI neither built nor formatted it, and it was pinned to adk-kotlin 0.2.0 — seven
minor versions behind. Its web entry point used AdkWebServer, which 0.9.0
deprecated and 1.0 removes. The frozen-path policy exempts deletions and
renames precisely so a recipe can be moved out, so this moves it.

It lands in core/kotlin rather than contrib/kotlin. contrib is the community
tier — its recipes are owned by individuals and outside organisations — while
core holds recipes curated by first-party teams, and this is one. core requires
an AGENTS.md, which this adds.

WebMain.kt now builds an AdkDevServer from an AdkServerConfig. AdkServerConfig
.inMemory() supplies the agent loader and the in-memory session and artifact
services the old constructor took one by one, so three imports go. The old code
passed ApiServerSpanExporter() explicitly; the new constructor defaults it to
the same thing, so the Dev UI trace view still works. The new classes bind
loopback where AdkWebServer pinned 0.0.0.0, so the README says so and names the
host setting in prose. It deliberately prints no worked example: the only value
such a snippet could carry is either 127.0.0.1, which overrides nothing, or
0.0.0.0, which is a copyable way to publish an unauthenticated server on every
interface. The reader is told to bind the narrowest one that works instead. The
startup log reads host and port off the config rather than repeating them as a
literal, so the one edit the README invites does not leave it announcing an
address the server is not listening on.

The three adk-kotlin coordinates now read a single `adkVersion` val rather
than repeating the literal. They have to move together — the KSP processor
generates against core's API — and a bump that edits two of three would
resolve a mismatched pair.

One doc correction the move forces: the Kotlin page of the recipe handbook
opened "Not currently active. No Kotlin recipes have landed yet", which stops
being true here, and kotlin/README.md links contributors straight to it. Only
that paragraph changes. The same sentence also claimed no Kotlin CI exists,
which has been stale since the format and test workflows landed in August and
is not this change's to fix; it goes only because it shares the sentence.

kotlin/agents is now empty, so kotlin/README.md points at the two live roots
instead. Its "browse agents/" instructions pointed at nothing, and it carried a
dead ../android/ link.

Verified: the recipe compiles against published 0.9.0 via `gradle test`, the
exact command the Kotlin test workflow runs; ktlint 1.8.0 (the pinned version,
checksum-matched) is clean, which it was not before — the four pre-existing
files had ten findings that had never been checked, since kotlin/agents was
excluded from that workflow; `uv run validate` passes all four checks, and
removing AGENTS.md was confirmed to fail structural validation, so the core
requirement is genuinely met. The recipe has no tests, so `gradle test` proves
compilation, not behaviour.
