+++ core/kotlin/llm-auditor/manifest.yaml
@@ -1,0 +1,30 @@
+type: "standalone"    # Options: [standalone | module]
+status: "active"      # Options: [active | inactive]
+language: "kotlin"    # Options: [python | java | go | kotlin | typescript]
+description: "A sequential multi-agent pipeline that fact-checks LLM-generated answers: a critic verifies each claim with Google Search, then a reviser minimally edits the answer to correct inaccuracies."
+
+architecture:
+  agent: "multi"       # Options: [single | multi]
+  stateful: false      # Options: [true | false]
+  datasources:         # Options: [hardcoded | local | external]
+    - "external"
+
+dependencies:
+  libraries:
+    - "ADK"
+  services:
+    - "Gemini API"
+    - "Google Search"
+
+ownership:
+  team: "adk-kotlin"
+  poc: "sherryfox"
+
+tags:
+  - "multi-agent"
+  - "sequential-agent"
+  - "fact-checking"
+  - "google-search"
+  - "callbacks"
+
+license: "Apache-2.0"
