+++ core/kotlin/llm-auditor/settings.gradle.kts
