+++ core/kotlin/llm-auditor/src/main/kotlin/com/google/adk/samples/agents/llmauditor/CriticPrompt.kt
