+++ core/kotlin/llm-auditor/src/main/kotlin/com/google/adk/samples/agents/llmauditor/WebMain.kt
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2026 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.adk.samples.agents.llmauditor
+
+import com.google.adk.kt.webserver.AdkServerConfig
+import com.google.adk.kt.webserver.dev.AdkDevServer
+
+fun main() {
+    // inMemory() supplies the agent loader and the session and artifact
+    // services, holding their state in the process.
+    val config = AdkServerConfig.inMemory(LlmAuditorAgent.rootAgent)
+
+    println("Starting ADK dev server on http://${config.host}:${config.port}")
+    AdkDevServer(config).start(wait = true)
+}
