+++ docs/recipe-handbook/languages/kotlin.md
@@ -1,10 +1,9 @@
-<!-- word count: 105 (target 100, cap 200) -->
+<!-- word count: 101 (target 100, cap 200) -->
 
 # Kotlin Recipes
 
-Not currently active. No Kotlin recipes have landed yet and no
-language-specific tooling (skills, CI, `pyproject`-equivalent
-checks) is in place.
+`core/kotlin/llm-auditor` is the first Kotlin recipe. There are no
+Kotlin authoring skills or `pyproject`-equivalent checks yet.
 
 **If you want to contribute one:** open a GitHub issue at
 [github.com/google/adk-samples/issues](https://github.com/google/adk-samples/issues)
